Consolidate content node utilities into kolibri-common.
