+++ packages/kolibri-common/utils/contentNode.js
@@ -1,5 +1,6 @@
 import groupBy from 'lodash/groupBy';
 import sortBy from 'lodash/sortBy';
+import { MasteryModelTypes } from 'kolibri.coreVue.vuex.constants';
 import { getContentLangActive } from 'kolibri.utils.i18n';
 
 export function deduplicateResources(contentNodes) {
@@ -29,4 +30,22 @@ export function deduplicateResources(contentNodes) {
     primaryNode.copies = sortedNodes;
     return primaryNode;
   });
+}
+
+export function masteryModelValidator({ type, m, n }) {
+  let isValid = true;
+  const typeIsValid = Object.values(MasteryModelTypes).includes(type);
+  if (!typeIsValid) {
+    // eslint-disable-next-line no-console
+    console.error(`Invalid mastery model type: ${type}`);
+    isValid = false;
+  }
+  if (type === MasteryModelTypes.m_of_n) {
+    if (typeof n !== 'number' || typeof m !== 'number') {
+      // eslint-disable-next-line no-console
+      console.error(`Invalid value of m and/or n. m: ${m}, n: ${n}`);
+      isValid = false;
+    }
+  }
+  return isValid;
 }
